requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
python-slugify==8.0.1
pillow==10.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
jinja2==3.1.2
markdown2==2.4.10
pygments==2.17.2 